# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workflows": {
      "post": {
        "description": "Validates the cron expression, timezone and every step config, saves the workflow, and registers its cron schedule. New workflows start `ACTIVE`. If the scheduler (Redis) is unavailable nothing is saved.",
        "operationId": "Workflows_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowDto"
              },
              "examples": {
                "httpPipeline": {
                  "summary": "HTTP → delay → HTTP",
                  "value": {
                    "name": "Nightly data sync",
                    "description": "Warms the reporting cache, waits, then triggers the export",
                    "cronExpression": "0 2 * * *",
                    "timezone": "Europe/London",
                    "steps": [
                      {
                        "stepOrder": 1,
                        "type": "HTTP",
                        "config": {
                          "url": "https://httpbin.org/get",
                          "method": "GET"
                        }
                      },
                      {
                        "stepOrder": 2,
                        "type": "DELAY",
                        "config": {
                          "duration": 2000
                        }
                      },
                      {
                        "stepOrder": 3,
                        "type": "HTTP",
                        "config": {
                          "url": "https://httpbin.org/post",
                          "method": "POST",
                          "headers": {
                            "X-Source": "cronflow"
                          },
                          "body": {
                            "report": "daily"
                          }
                        },
                        "retryCount": 5,
                        "timeout": 10000
                      }
                    ]
                  }
                },
                "minimal": {
                  "summary": "Single health ping every 5 minutes",
                  "value": {
                    "name": "Uptime ping",
                    "cronExpression": "*/5 * * * *",
                    "steps": [
                      {
                        "stepOrder": 1,
                        "type": "HTTP",
                        "config": {
                          "url": "https://httpbin.org/status/200"
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — e.g. invalid cron, duplicate stepOrder, HTTP step without url",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "Scheduler (Redis) unavailable — the workflow was not created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Create a workflow",
        "tags": [
          "Workflows"
        ]
      },
      "get": {
        "description": "Paginated list of your workflows, filterable by status and a text search.",
        "operationId": "Workflows_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Only return workflows with this status",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "PAUSED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive match against name or description",
            "schema": {
              "maxLength": 100,
              "example": "sync",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt",
                "name"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedWorkflowsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "List workflows",
        "tags": [
          "Workflows"
        ]
      }
    },
    "/api/v1/workflows/{id}": {
      "get": {
        "description": "Includes its steps and webhook token.",
        "operationId": "Workflows_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Workflow id",
            "schema": {
              "example": "cm0x8b1f40001abcdlkj2h3g4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Get a workflow",
        "tags": [
          "Workflows"
        ]
      },
      "patch": {
        "description": "Partial update. Sending `steps` replaces the entire step list (past executions keep their history). Changing `cronExpression` or `timezone` reschedules an active workflow immediately.",
        "operationId": "Workflows_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Workflow id",
            "schema": {
              "example": "cm0x8b1f40001abcdlkj2h3g4",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, or the body contains no fields to update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (PostgreSQL or Redis) is temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Update a workflow",
        "tags": [
          "Workflows"
        ]
      },
      "delete": {
        "description": "Removes its schedule and all of its execution history.",
        "operationId": "Workflows_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Workflow id",
            "schema": {
              "example": "cm0x8b1f40001abcdlkj2h3g4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workflow deleted"
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (PostgreSQL or Redis) is temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Delete a workflow",
        "tags": [
          "Workflows"
        ]
      }
    },
    "/api/v1/workflows/{id}/pause": {
      "patch": {
        "description": "Stops scheduled and webhook runs. Manual runs are still allowed.",
        "operationId": "Workflows_pause",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Workflow id",
            "schema": {
              "example": "cm0x8b1f40001abcdlkj2h3g4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Workflow is already paused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (PostgreSQL or Redis) is temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Pause a workflow",
        "tags": [
          "Workflows"
        ]
      }
    },
    "/api/v1/workflows/{id}/resume": {
      "patch": {
        "description": "Re-registers its cron schedule.",
        "operationId": "Workflows_resume",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Workflow id",
            "schema": {
              "example": "cm0x8b1f40001abcdlkj2h3g4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Workflow is already active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (PostgreSQL or Redis) is temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Resume a paused workflow",
        "tags": [
          "Workflows"
        ]
      }
    },
...
  },
  "info": {
    "title": "CronFlow API",
    "description": "CronFlow runs multi-step workflows — HTTP calls and delays — on a **cron schedule**, from a **webhook**, or **on demand**.\n\n### Quick start\n1. `POST /api/v1/auth/register` and copy the `accessToken`\n2. Click **Authorize**, paste the token into **JWT**\n3. `POST /api/v1/workflows` to create a workflow (see the request examples)\n4. `POST /api/v1/workflows/{workflowId}/executions` to run it now, then poll `GET /api/v1/executions/{id}`\n\n### Authentication\n| Method | Header | Obtain from |\n| --- | --- | --- |\n| JWT | `Authorization: Bearer <accessToken>` | `POST /api/v1/auth/login` |\n| API key | `x-api-key: cf_…` | `POST /api/v1/api-keys` |\n\n### Errors\nEvery error response uses the same envelope (`ErrorResponseDto`). Its `requestId` matches the\n`X-Request-Id` response header and the server logs.\n\n### Rate limiting\nRequests are limited per client IP (100/minute by default; 10/minute for register and login).\nExceeding the limit returns `429 Too Many Requests`.",
    "version": "3.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token from POST /api/v1/auth/login"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key from POST /api/v1/api-keys"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "HTTP reason phrase"
          },
          "message": {
            "description": "A single message, or one message per failed validation rule",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "cronExpression must be a valid 5-field cron expression (minute hour day-of-month month day-of-week)"
            ]
          },
          "path": {
            "type": "string",
            "example": "/api/v1/workflows"
          },
          "method": {
            "type": "string",
            "example": "POST"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-14T10:00:00.000Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "5f0e7c1a-3b8e-4d7a-9c2f-1e6b8a4d2c10",
            "description": "Correlates this response with server logs (also sent as the X-Request-Id header)"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "method",
          "timestamp"
        ]
      },
...
      "CreateWorkflowDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Nightly data sync",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "example": "Warms the reporting cache, waits, then triggers the export",
            "maxLength": 500
          },
          "cronExpression": {
            "type": "string",
            "example": "0 2 * * *",
            "description": "Standard 5-field cron expression: minute hour day-of-month month day-of-week"
          },
          "timezone": {
            "type": "string",
            "example": "Europe/London",
            "default": "UTC",
            "description": "IANA timezone the cron expression is evaluated in"
          },
          "steps": {
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStepDto"
            }
          }
        },
        "required": [
          "name",
          "cronExpression",
          "steps"
        ]
      },
...
      "WorkflowResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm0x8b1f40001abcdlkj2h3g4"
          },
          "name": {
            "type": "string",
            "example": "Nightly data sync"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Warms the reporting cache, waits, then triggers the export"
          },
          "cronExpression": {
            "type": "string",
            "example": "0 2 * * *"
          },
          "timezone": {
            "type": "string",
            "example": "UTC"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PAUSED"
            ],
            "example": "ACTIVE"
          },
          "nextRunAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-09-15T02:00:00.000Z",
            "description": "Next scheduled run; null while the workflow is paused"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-14T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-14T10:00:00.000Z"
          },
          "webhookToken": {
            "type": "string",
            "example": "k3Jd9sQ2mVx7LpA0bR4tYw8eZc1uN6hG",
            "description": "Secret for `POST /api/v1/webhooks/{token}/trigger`. Treat it like a password."
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStepResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "cronExpression",
          "timezone",
          "status",
          "nextRunAt",
          "createdAt",
          "updatedAt",
          "webhookToken",
          "steps"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "PaginatedWorkflowsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateWorkflowDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Nightly data sync",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "example": "Warms the reporting cache, waits, then triggers the export",
            "maxLength": 500
          },
          "cronExpression": {
            "type": "string",
            "example": "0 2 * * *",
            "description": "Standard 5-field cron expression: minute hour day-of-month month day-of-week"
          },
          "timezone": {
            "type": "string",
            "example": "Europe/London",
            "default": "UTC",
            "description": "IANA timezone the cron expression is evaluated in"
          },
          "steps": {
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStepDto"
            }
          }
        }
      },
...
    }
  }
}